application refactorings
