+++ app/src/main/java/com/example/weatherprovider/location/LocationDAO.kt
@@ -1,4 +1,4 @@
-package com.example.weatherprovider.model
+package com.example.weatherprovider.location
 
 import androidx.lifecycle.LiveData
 import androidx.room.*
