+++ app/src/main/res/layout/activity_city_search.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CitySearchActivity">
+    tools:context=".citySearch.CitySearchActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
