+++ app/src/main/res/layout/fragment_location_forecast.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".forecast.ForecastFragment">
+    tools:context=".location.LocationFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -12,8 +12,8 @@
 
         <TextView
             android:id="@+id/woeid_text"
-            android:layout_width="126dp"
-            android:layout_height="27dp"
+            android:layout_width="60dp"
+            android:layout_height="20dp"
             android:text="TextView"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
